thesis/requirements.txt
django>=4.0.*
python>=3.9.*
django_plotly_dash0>=1.6.*
channels>=3.0.*
channels_redis>=3.3.* 
crispy-forms>=1.4.*
djangorestframework>=3.13.*
geocoder>=1.38.*
folium>=0.12.* 
django-cors-headers>=3.11.*
